backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
spacy
chromadb
google-genai
networkx
sqlalchemy
pydantic
python-dotenv
httpx
pymongo
Pillow
